fix: merge issues with `Coordinate2D`
