+++ include/mcpp/coordinate.h
@@ -7,6 +7,9 @@
  *
  */
 namespace mcpp {
+
+struct Coordinate2D;
+
 /**
  * Represented using integers since sub-unit coordinates are not of particular
  * relevance. Allows for operations such as addition between coordinates.
@@ -48,6 +51,8 @@ struct Coordinate {
    * coordinates.
    */
   Coordinate operator+(const Coordinate& obj) const;
+
+  Coordinate operator+(const Coordinate2D& obj) const;
 
   /**
    * @brief Checks if two Coordinate objects are equal.
@@ -91,6 +96,97 @@ struct Coordinate {
    * @return The output stream with the Coordinate object's values.
    */
   friend std::ostream& operator<<(std::ostream& out, const Coordinate& coord);
+};
+
+/**
+ * @brief Height-agnostic coordinate class.
+ *
+ * Represented using integers since sub-unit coordinates are not of particular
+ * relevance. Allows for operations such as addition between flat coordinates.
+ */
+struct Coordinate2D {
+  /**
+   * @brief Constructs a Coordinate2D object with integer values.
+   *
+   * @param x The x-coordinate.
+   * @param z The z-coordinate.
+   */
+  constexpr Coordinate2D(int x, int z) : x(x), z(z) {}
+
+  /**
+   * @brief Constructs a Coordinate2D object with zero values.
+   */
+  constexpr Coordinate2D() : x(0), z(0) {}
+
+  /**
+   * @brief Constructs a Coordinate2D object with double values.
+   *
+   * @param x The x-coordinate as a double.
+   * @param z The z-coordinate as a double.
+   */
+  constexpr Coordinate2D(double x, double z) : x(static_cast<int>(x)), z(static_cast<int>(z)) {}
+
+  /**
+   * @brief Constructs a Coordinate2D object from a Coordinate object.
+   *
+   * @param coord The Coordinate object.
+   */
+  constexpr Coordinate2D(const Coordinate& coord) : x(coord.x), z(coord.z) {}
+
+  /**
+   * @brief Constructs a Coordinate object from a Coordinate2D object and a
+   * y value.
+   *
+   * @param coord The Coordinate2D object.
+   * @param y The y value.
+   */
+  Coordinate with_height(int y) const;
+
+  /**
+   * @brief Adds two Coordinate2D objects.
+   *
+   * @param obj The Coordinate2D object to add.
+   * @return A new Coordinate2D object representing the sum of the two
+   * coordinates.
+   */
+  Coordinate2D operator+(const Coordinate2D& obj) const;
+
+  /**
+   * @brief Checks if two Coordinate2D objects are equal.
+   *
+   * @param obj The Coordinate2D object to compare with.
+   * @return True if the flat coordinates are equal, false otherwise.
+   */
+  bool operator==(const Coordinate2D& obj) const;
+
+  /**
+   * @brief Checks if two Coordinate2D objects are not equal.
+   *
+   * @param obj The Coordinate2D object to compare with.
+   * @return True if the flat coordinates are not equal, false otherwise.
+   */
+  bool operator!=(const Coordinate2D& obj) const;
+
+  /**
+   * @brief Subtracts one Coordinate2D object from another.
+   *
+   * @param obj The Coordinate2D object to subtract.
+   * @return A new Coordinate2D object representing the difference between
+   * the two coordinates.
+   */
+  Coordinate2D operator-(const Coordinate2D& obj) const;
+
+  /**
+   * @brief Outputs the Coordinate2D object to an ostream.
+   *
+   * @param out The output stream.
+   * @param coord The Coordinate2D object to output.
+   * @return The output stream with the Coordinate object's values.
+   */
+  friend std::ostream& operator<<(std::ostream& out, const Coordinate2D& coord);
+
+  int x;
+  int z;
 };
 
 /**
